backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
Pillow==10.0.0
djoser==2.2.0
drf-extra-fields==3.7.0
django-filter==23.2
reportlab==4.0.4